fix(live-window): address code review feedback for stars utility
